+++ types.go
@@ -1,6 +1,6 @@
 package exception
 
 type Exception struct {
-	Code int    `json:"code"`
+	Code int64  `json:"code"`
 	Msg  string `json:"msg"`
 }
